@@ -1,116 +1,153 @@
 package org.vaccineimpact.orderlyweb.tests.security
 
 import com.nhaarman.mockito_kotlin.*
-import org.assertj.core.api.Assertions
-import org.eclipse.egit.github.core.Team
-import org.eclipse.egit.github.core.User
-import org.eclipse.egit.github.core.client.GitHubClient
-import org.eclipse.egit.github.core.client.GitHubResponse
-import org.eclipse.egit.github.core.client.RequestException
-import org.junit.Assert
+import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.Test
 import org.pac4j.core.exception.CredentialsException
+import org.kohsuke.github.*
 import org.vaccineimpact.orderlyweb.db.Config
+import org.vaccineimpact.orderlyweb.db.InvalidConfigurationKey
 import org.vaccineimpact.orderlyweb.errors.BadConfigurationError
 import org.vaccineimpact.orderlyweb.security.providers.GithubApiClientAuthHelper
 
 class GithubApiClientAuthHelperTests
 {
-    private val mockTeam = Team().apply {
-        name = "teamName"
-        id = 1
+    private val token = "token"
+    private val orgName = "orgName"
+    private val teamName = "teamName"
+
+    private val orgNameWithoutUser = "orgNameWithoutUser"
+    private val teamNameWithoutUser = "teamNameWithoutUser"
+
+    private val mockTeam = mock<GHTeam>()
+    private val mockTeamWithoutUser = mock<GHTeam>()
+
+    private val mockOrg = mock<GHOrganization>{
+        on { name } doReturn orgName
+        on { teams } doReturn mapOf(teamName to mockTeam, teamNameWithoutUser to mockTeamWithoutUser)
     }
 
-    private val mockUser = User().apply {
-        login = "user.name"
-        email = null
-        name = "full name"
+    private val mockOrgWithoutUser = mock<GHOrganization>()
+
+    private val mockEmail = mock<GHEmail>{
+        on { email } doReturn "privateEmail"
     }
 
-    private val mockOrg = User().apply {
-        login = "orgName"
+    private val mockUser = mock<GHMyself>{
+        on { isMemberOf(mockOrg) } doReturn true
+        on { isMemberOf(mockTeam) } doReturn true
+        on { isMemberOf(mockOrgWithoutUser) } doReturn false
+        on { isMemberOf(mockTeamWithoutUser) } doReturn false
+        on { emails2 } doReturn listOf(mockEmail)
     }
+
+    private val mockGithub = mock<GitHub> {
+        on { myself } doReturn mockUser
+        on { getOrganization(orgName) } doReturn mockOrg
+        on { getOrganization(orgNameWithoutUser) } doReturn mockOrgWithoutUser
+    }
+
+    private fun getGitHubBuilder(github: GitHub): GitHubBuilder
+    {
+        val mockBuilderResult = mock<GitHubBuilder> {
+            on { build() } doReturn github
+        }
+
+        return mock {
+            on { withOAuthToken(token) } doReturn mockBuilderResult
+        }
+    }
+
+    private val mockGithubBuilder = getGitHubBuilder(mockGithub)
 
     private val mockAppConfig = mock<Config> {
-        on { get("auth.github_org") } doReturn "orgName"
-        on { get("auth.github_team") } doReturn "teamName"
-    }
-
-    private val mockGithubApiClient = mock<GitHubClient> {
-        on { get(argWhere { it.uri.contains("/user/emails") }) } doReturn
-                GitHubResponse(mock(), listOf("privateEmail"))
-        on { get(argWhere { it.uri.endsWith("/user") }) } doReturn
-                GitHubResponse(mock(), mockUser)
-        on { get(argWhere { it.uri.contains("user/orgs") }) } doReturn
-                GitHubResponse(mock(), listOf(mockOrg))
-        on { get(argWhere { it.uri.contains("orgs/orgName/teams") }) } doReturn
-                GitHubResponse(mock(), listOf(mockTeam))
-        on { get(argWhere { it.uri.contains("teams/1/members") }) } doReturn
-                GitHubResponse(mock(), listOf(mockUser))
-
+        on { get("auth.github_org") } doReturn orgName
+        on { get("auth.github_team") } doReturn teamName
     }
 
     @Test
-    fun `initialise fails if token is blank`()
+    fun `authenticate fails if token is blank`()
     {
-        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubApiClient)
+        val sut = GithubApiClientAuthHelper(mock(), mock())
 
-        Assertions.assertThatThrownBy { sut.authenticate("") }
+        assertThatThrownBy { sut.authenticate("") }
                 .isInstanceOf(CredentialsException::class.java)
                 .hasMessageContaining("Token cannot be blank")
     }
 
     @Test
-    fun `can initialise`()
+    fun `can authenticate`()
     {
-        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubApiClient)
-        sut.authenticate("token")
-        verify(mockGithubApiClient).setOAuth2Token("token")
+        val ghBuilder = mockGithubBuilder
+        val sut = GithubApiClientAuthHelper(mockAppConfig, ghBuilder)
+        sut.authenticate(token)
+        verify(ghBuilder).withOAuthToken(token)
     }
 
     @Test
     fun `checkGithubUserCanAuthenticate succeeds when user is member of allowed org and team`()
     {
         //default mock values should succeed
-        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubApiClient)
-        sut.authenticate("token")
+        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubBuilder)
+        sut.authenticate(token)
+        sut.checkGitHubOrgAndTeamMembership()
+    }
+
+    @Test
+    fun `checkGithubUserCanAuthenticate succeeds when user is member of allowed org and team is not configured`()
+    {
+        val customMockAppConfig = mock<Config> {
+            on { get("auth.github_org") } doReturn orgName
+            on { get("auth.github_team") } doReturn ""
+        }
+
+        val sut = GithubApiClientAuthHelper(customMockAppConfig, mockGithubBuilder)
+        sut.authenticate(token)
         sut.checkGitHubOrgAndTeamMembership()
     }
 
     @Test
     fun `can getUser`()
     {
-        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubApiClient)
-        sut.authenticate("token")
+        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubBuilder)
+        sut.authenticate(token)
         val result = sut.getUser()
-        Assert.assertSame(mockUser, result)
-    }
-
-    @Test
-    fun `can getUserEmail when email is public`()
-    {
-        mockUser.email = "publicEmail"
-        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubApiClient)
-        sut.authenticate("token")
-        val result = sut.getUserEmail()
-        Assert.assertEquals("publicEmail", result)
+        assertThat(mockUser).isSameAs(result)
     }
 
     @Test
     fun `can getUserEmail when email is private`()
     {
         //defaults should get the sut to find null email in user so fetch details from client
-        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubApiClient)
-        sut.authenticate("token")
+        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubBuilder)
+        sut.authenticate(token)
         val result = sut.getUserEmail()
-        Assert.assertEquals("privateEmail", result)
+        assertThat(result).isEqualTo("privateEmail")
     }
 
     @Test
-    fun `on checkGithubUserHasOrderlyWebAccess, IllegalStateException thrown if not authenticated`()
+    fun `can getUserEmail when email is public`()
     {
-        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubApiClient)
-        Assertions.assertThatThrownBy {
+        val customMockUser = mock<GHMyself>{
+            on { email } doReturn "publicEmail"
+        }
+
+        val customMockGithub = mock<GitHub> {
+            on { myself } doReturn customMockUser
+        }
+
+        val sut = GithubApiClientAuthHelper(mockAppConfig, getGitHubBuilder(customMockGithub))
+        sut.authenticate(token)
+        val result = sut.getUserEmail()
+        assertThat(result).isEqualTo("publicEmail")
+    }
+
+    @Test
+    fun `on checkGithubOrgAndTeamMembership IllegalStateException thrown if not authenticated`()
+    {
+        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubBuilder)
+        assertThatThrownBy {
             sut.checkGitHubOrgAndTeamMembership()
         }.isInstanceOf(IllegalStateException::class.java)
                 .hasMessageContaining("User has not been authenticated")
@@ -119,8 +156,8 @@ class GithubApiClientAuthHelperTests
     @Test
     fun `on getUser, IllegalStateException thrown if not authenticated`()
     {
-        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubApiClient)
-        Assertions.assertThatThrownBy {
+        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubBuilder)
+        assertThatThrownBy {
             sut.getUser()
         }.isInstanceOf(IllegalStateException::class.java)
                 .hasMessageContaining("User has not been authenticated")
@@ -129,83 +166,62 @@ class GithubApiClientAuthHelperTests
     @Test
     fun `on getUserEmail, IllegalStateException thrown if not authenticated`()
     {
-        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubApiClient)
-        Assertions.assertThatThrownBy {
+        val sut = GithubApiClientAuthHelper(mockAppConfig, mockGithubBuilder)
+        assertThatThrownBy {
             sut.getUserEmail()
         }.isInstanceOf(IllegalStateException::class.java)
                 .hasMessageContaining("User has not been authenticated")
     }
 
     @Test
-    fun `on check access, BadConfigurationError is thrown if team does not exist`()
+    fun `on check team membership, BadConfigurationError is thrown if team is not in org`()
     {
-        val customMockGithubApiClient = mock<GitHubClient> {
-            on { get(argWhere { it.uri.contains("user") }) } doReturn
-                    GitHubResponse(mock(), mockUser)
-            on { get(argWhere { it.uri.contains("user/orgs") }) } doReturn
-                    GitHubResponse(mock(), listOf(mockOrg))
-            on { get(argWhere { it.uri.endsWith("orgs/orgName/teams") }) } doReturn GitHubResponse(mock(),
-                    listOf<Team>())
+        val customMockAppConfig = mock<Config> {
+            on { get("auth.github_org") } doReturn orgName
+            on { get("auth.github_team") } doReturn "teamNotInOrg"
         }
 
-        val mockAppConfig = mock<Config> {
-            on { get("auth.github_org") } doReturn "orgName"
-            on { get("auth.github_team") } doReturn "teamName"
-        }
-
-        val sut = GithubApiClientAuthHelper(mockAppConfig, customMockGithubApiClient)
+        val sut = GithubApiClientAuthHelper(customMockAppConfig, mockGithubBuilder)
 
         sut.authenticate("token")
 
-        Assertions.assertThatThrownBy {
+        assertThatThrownBy {
             sut.checkGitHubOrgAndTeamMembership()
         }.isInstanceOf(BadConfigurationError::class.java)
-                .hasMessageContaining("GitHub org orgName has no team called teamName")
+                .hasMessageContaining("GitHub org orgName has no team called teamNotInOrg")
     }
 
     @Test
-    fun `on check access, CredentialsException is thrown if user does not belong to GitHub org`()
+    fun `on check org membership, CredentialsException is thrown if user does not belong to GitHub org`()
     {
-        val customMockGithubApiClient = mock<GitHubClient> {
-            on { get(argWhere { it.uri.endsWith("user") }) } doReturn
-                    GitHubResponse(mock(), mockUser)
-            on { get(argWhere { it.uri.contains("user/orgs") } ) } doReturn GitHubResponse(mock(),
-                    listOf<User>())
+        val customMockAppConfig = mock<Config> {
+            on { get("auth.github_org") } doReturn orgNameWithoutUser
+            on { get("auth.github_team") } doReturn teamName
         }
 
-        val mockAppConfig = mock<Config> {
-            on { get("auth.github_org") } doReturn "orgName"
-        }
-
-        val sut = GithubApiClientAuthHelper(mockAppConfig, customMockGithubApiClient)
+        val sut = GithubApiClientAuthHelper(customMockAppConfig, mockGithubBuilder)
 
         sut.authenticate("token")
 
-        Assertions.assertThatThrownBy {
+        assertThatThrownBy {
             sut.checkGitHubOrgAndTeamMembership()
         }.isInstanceOf(CredentialsException::class.java)
                 .hasMessageContaining("User is not a member of GitHub org orgName")
     }
 
     @Test
-    fun `on check access, CredentialsException is thrown if user does not belong to GitHub team`()
+    fun `on check team membership, CredentialsException is thrown if user does not belong to GitHub team`()
     {
-        val customMockGithubApiClient = mock<GitHubClient> {
-            on { get(argWhere { it.uri.endsWith("user") }) } doReturn
-                    GitHubResponse(mock(), mockUser)
-            on { get(argWhere { it.uri.contains("user/orgs") }) } doReturn
-                    GitHubResponse(mock(), listOf(mockOrg))
-            on { get(argWhere { it.uri.contains("orgs/orgName/teams") }) } doReturn
-                    GitHubResponse(mock(), listOf(mockTeam))
-            on { get(argWhere { it.uri.contains("teams/1/members") }) } doReturn
-                    GitHubResponse(mock(), listOf<User>())
+        val customMockAppConfig = mock<Config> {
+            on { get("auth.github_org") } doReturn orgName
+            on { get("auth.github_team") } doReturn teamNameWithoutUser
         }
 
-        val sut = GithubApiClientAuthHelper(mockAppConfig, customMockGithubApiClient)
+        val sut = GithubApiClientAuthHelper(customMockAppConfig, mockGithubBuilder)
 
         sut.authenticate("token")
 
-        Assertions.assertThatThrownBy {
+        assertThatThrownBy {
             sut.checkGitHubOrgAndTeamMembership()
         }.isInstanceOf(CredentialsException::class.java)
                 .hasMessageContaining("User is not a member of GitHub team teamName")
@@ -214,47 +230,126 @@ class GithubApiClientAuthHelperTests
     @Test
     fun `on getUserEmail, CredentialsException is thrown if email is not public and token does not have email scope`()
     {
-        val userWithNullEmail = mockUser.apply { email = null }
-        val customMockGithubApiClient = mock<GitHubClient> {
-            on { get(argWhere { it.uri.endsWith("user") }) } doReturn
-                    GitHubResponse(mock(), userWithNullEmail)
-            on { get(argWhere { it.uri.contains("user/orgs") }) } doReturn
-                    GitHubResponse(mock(), listOf(mockOrg))
-            on { get(argWhere { it.uri.contains("orgs/orgName/teams") }) } doReturn
-                    GitHubResponse(mock(), listOf(mockTeam))
-            on { get(argWhere { it.uri.contains("teams/1/members") }) } doReturn
-                    GitHubResponse(mock(), listOf(userWithNullEmail))
-            on { get(argWhere { it.uri.contains("emails") }) } doThrow RequestException(mock(), 404)
+        val customMockUser = mock<GHMyself>{
+            on { emails2 } doThrow GHFileNotFoundException()
         }
 
-        val sut = GithubApiClientAuthHelper(mockAppConfig, customMockGithubApiClient)
+        val customMockGithub = mock<GitHub> {
+            on { myself } doReturn customMockUser
+        }
+
+        val sut = GithubApiClientAuthHelper(mockAppConfig, getGitHubBuilder(customMockGithub))
 
         sut.authenticate("token")
 
-        Assertions.assertThatThrownBy {
+        assertThatThrownBy {
             sut.getUserEmail()
         }.isInstanceOf(CredentialsException::class.java)
                 .hasMessageContaining("GitHub token must include scope user:email")
     }
 
     @Test
-    fun `on check access, Exception is thrown if org membership is required and token does not have read user scope`()
+    fun `throws CredentialsException if unauthorized when getting user`()
     {
-        val customMockGithubApiClient = mock<GitHubClient> {
-            on { get(argWhere { it.uri.contains("user") }) } doReturn
-                    GitHubResponse(mock(), mockUser)
-            on { get(argWhere { it.uri.contains("user/orgs") }) } doThrow RequestException(mock(), 403)
+        val customMockGithub = mock<GitHub> {
+            on { myself } doThrow HttpException("Test error", 401, "", "")
         }
 
-        val sut = GithubApiClientAuthHelper(mockAppConfig, customMockGithubApiClient)
-
-        sut.authenticate("token")
-
-        Assertions.assertThatThrownBy {
-            sut.checkGitHubOrgAndTeamMembership()
+        val sut = GithubApiClientAuthHelper(mockAppConfig, getGitHubBuilder(customMockGithub))
+        assertThatThrownBy {
+            sut.authenticate(token)
         }.isInstanceOf(CredentialsException::class.java)
-                .hasMessageContaining("GitHub token must include scope read:user")
-
+                .hasMessage("Test error")
     }
 
+    @Test
+    fun `throws CredentialsException if unauthorized when getting user with null message`()
+    {
+        val customMockGithub = mock<GitHub> {
+            on { myself } doThrow HttpException(null, 401, "", "")
+        }
+
+        val sut = GithubApiClientAuthHelper(mockAppConfig, getGitHubBuilder(customMockGithub))
+        assertThatThrownBy {
+            sut.authenticate(token)
+        }.isInstanceOf(CredentialsException::class.java)
+                .hasMessage("")
+    }
+
+    @Test
+    fun `throws non-401 HttpException when getting user`()
+    {
+        val httpException = HttpException("TestError", 500, "", "")
+        val customMockGithub = mock<GitHub> {
+            on { myself } doThrow httpException
+        }
+
+        val sut = GithubApiClientAuthHelper(mockAppConfig, getGitHubBuilder(customMockGithub))
+        assertThatThrownBy {
+            sut.authenticate(token)
+        }.isSameAs(httpException)
+    }
+
+    @Test
+    fun `throws CredentialsException if unauthorized when getting org`()
+    {
+        val customMockGithub = mock<GitHub> {
+            on { myself } doReturn mockUser
+            on { getOrganization(orgName) } doThrow HttpException("Test error", 401, "", "")
+        }
+
+        val sut = GithubApiClientAuthHelper(mockAppConfig, getGitHubBuilder(customMockGithub))
+        sut.authenticate(token)
+        assertThatThrownBy {
+            sut.checkGitHubOrgAndTeamMembership()
+        }.isInstanceOf(CredentialsException::class.java)
+                .hasMessage("Test error")
+    }
+
+    @Test
+    fun `throws CredentialsException if unauthorized when getting org with null message`()
+    {
+        val customMockGithub = mock<GitHub> {
+            on { myself } doReturn mockUser
+            on { getOrganization(orgName) } doThrow HttpException(null, 401, "", "")
+        }
+
+        val sut = GithubApiClientAuthHelper(mockAppConfig, getGitHubBuilder(customMockGithub))
+        sut.authenticate(token)
+        assertThatThrownBy {
+            sut.checkGitHubOrgAndTeamMembership()
+        }.isInstanceOf(CredentialsException::class.java)
+                .hasMessage("")
+    }
+
+    @Test
+    fun `throws non-401 HttpException when getting org`()
+    {
+        val httpException = HttpException("TestError", 500, "", "")
+        val customMockGithub = mock<GitHub> {
+            on { myself } doReturn mockUser
+            on { getOrganization(orgName) } doThrow httpException
+        }
+
+        val sut = GithubApiClientAuthHelper(mockAppConfig, getGitHubBuilder(customMockGithub))
+        sut.authenticate(token)
+        assertThatThrownBy {
+            sut.checkGitHubOrgAndTeamMembership()
+        }.isSameAs(httpException)
+    }
+
+    @Test
+    fun `throws InvalidConfigurationKey if configured org is empty`()
+    {
+        val customMockAppConfig = mock<Config> {
+            on { get("auth.github_org") } doReturn ""
+            on { get("auth.github_team") } doReturn ""
+        }
+        val sut = GithubApiClientAuthHelper(customMockAppConfig, mockGithubBuilder)
+        sut.authenticate(token)
+        assertThatThrownBy {
+            sut.checkGitHubOrgAndTeamMembership()
+        }.isInstanceOf(InvalidConfigurationKey::class.java)
+                .hasMessage("Invalid configuration value '' for key 'auth.github_org'")
+    }
 }
