@@ -3,6 +3,8 @@ import Vuex, {StoreOptions} from "vuex";
 import {git} from "../git/git";
 import {RunnerRootState} from "../../utils/types";
 import {mutations} from "./mutations";
+import {reports} from "../reports/reports";
+import {errors} from "../errors/errors";
 
 export type RunReportTabName = "RunReport" | "ReportLogs"
 
@@ -11,7 +13,9 @@ export interface RunReportRootState extends RunnerRootState {
 }
 
 export const namespace = {
-    git: "git"
+    git: "git",
+    reports: "reports",
+    errors: "errors"
 }
 
 export const storeOptions: StoreOptions<RunReportRootState> = {
@@ -19,7 +23,9 @@ export const storeOptions: StoreOptions<RunReportRootState> = {
         selectedTab: "RunReport"
     } as RunReportRootState,
     modules: {
-        [namespace.git]: git
+        [namespace.git]: git,
+        [namespace.reports]: reports,
+        [namespace.errors]: errors
     },
     mutations
 };
