@@ -1,4 +1,7 @@
 import {GitState} from "../store/git/git";
+import {ReportsState} from "../store/reports/reports";
+import {Payload} from "vuex";
+import {ErrorsState} from "../store/errors/errors";
 
 export interface Parameter {
     name: string,
@@ -20,7 +23,7 @@ export interface Report {
     id: string
 }
 
-export interface ReportLog{
+export interface ReportLog {
     email: string,
     date: string,
     report: string,
@@ -51,7 +54,7 @@ export interface WorkflowReportWithParams {
     params?: Record<string, string>
 }
 
-export interface Error {
+export interface ErrorResponse {
     response?: {
         data?: {
             errors?: any[]
@@ -132,11 +135,6 @@ export interface ChildCustomStyle {
     control: string
 }
 
-export interface Error {
-    code: string;
-    message: string | null;
-}
-
 export interface WorkflowReportWithDependencies {
     name: string,
     instance?: string
@@ -152,7 +150,36 @@ export interface WorkflowSummaryResponse {
 }
 
 export interface RunnerRootState {
-    git: GitState
+    git: GitState,
+    reports: ReportsState,
+    errors: ErrorsState
+}
+
+export interface GitCommit {
+    id: string,
+    date_time: string,
+    age:number
 }
 
 export type EmptyObject = Record<string, never>
+
+export type Dict<V> = { [k: string]: V }
+
+export type ComputedWithType<T> = () => T;
+
+export interface PayloadWithType<T> extends Payload {
+    payload: T
+}
+
+export interface Error {
+    code: string;
+    message: string | null;
+}
+
+export interface Response {
+    status: "success" | "failure";
+    data: {
+        [k: string]: any;
+    } | null;
+    errors: Error[];
+}
