@@ -1,42 +1,42 @@
-import {shallowMount} from "@vue/test-utils";
+import { shallowMount } from "@vue/test-utils";
 import runWorkflowSummary from "../../../../js/components/runWorkflow/workflowSummary/runWorkflowSummary.vue"
 import runWorkflowSummaryHeader from "../../../../js/components/runWorkflow/workflowSummary/runWorkflowSummaryHeader.vue"
-import {RunWorkflowMetadata, WorkflowSummaryResponse} from "../../../../js/utils/types";
-import {mockAxios} from "../../../mockAxios";
+import { RunWorkflowMetadata, WorkflowSummaryResponse } from "../../../../js/utils/types";
+import { mockAxios } from "../../../mockAxios";
 import workflowSummaryReports from "../../../../js/components/runWorkflow/workflowSummary/workflowSummaryReports.vue";
 
 describe(`runWorkflowSummary`, () => {
 
     const workflowSummary = {
         ref: "refNum",
-        missing_dependencies: {step2: ["step1"]},
-        reports: [{name: "step2", params: {} }]
+        missing_dependencies: { step2: ["step1"] },
+        reports: [{ name: "step2", params: {} }]
     }
 
     const workflowSummary2: WorkflowSummaryResponse = {
         ref: "test",
         missing_dependencies: {},
-        reports: [{name: "test", param_list: [{name: "key", value: "value"}]}]
+        reports: [{ name: "test", param_list: [{ name: "key", value: "value" }] }]
     }
 
     const metaData = {
-        reports: [{name: "r1"}],
+        reports: [{ name: "r1" }],
         git_commit: "gitCommit"
     }
 
     const metaData2 = {
-        reports: [{name: "r1", params: {"key": "value"}}, {name: "r2"}],
+        reports: [{ name: "r1", params: { "key": "value" } }, { name: "r2" }],
         git_commit: "gitCommit"
     }
 
     beforeEach(() => {
         mockAxios.reset();
         mockAxios.onPost('http://app/workflows/summary/?commit=gitCommit')
-            .reply(200, {"data": workflowSummary});
+            .reply(200, { "data": workflowSummary });
     })
 
-    const getWrapper = (meta: Partial<RunWorkflowMetadata> = {reports: []}) => {
-        return shallowMount(runWorkflowSummary, {propsData: {workflowMetadata: meta}})
+    const getWrapper = (meta: Partial<RunWorkflowMetadata> = { reports: [] }) => {
+        return shallowMount(runWorkflowSummary, { propsData: { workflowMetadata: meta } })
     }
 
     it(`it posts to workflow summary endpoint and renders workflow summary header`, (done) => {
@@ -51,8 +51,8 @@ describe(`runWorkflowSummary`, () => {
 
     it(`it can post workflow summary with dependencies as response`, (done) => {
         mockAxios.onPost('http://app/workflows/summary/?commit=gitCommit')
-            .reply(200, {"data": workflowSummary2});
-            
+            .reply(200, { "data": workflowSummary2 });
+
         const wrapper = getWrapper(metaData2);
 
         setTimeout(() => {
@@ -68,7 +68,7 @@ describe(`runWorkflowSummary`, () => {
 
     it(`emits valid event on mount`, () => {
         const wrapper = getWrapper({
-            reports: [{name: "r1"}]
+            reports: [{ name: "r1" }]
         });
         expect(wrapper.emitted().valid[0][0]).toBe(true);
     });
